refactor: rename session/ts/ to session/internal/
The "ts" name was meaningless — everything is TypeScript. "internal"
clearly signals these are implementation details behind the session API.
