+++ packages/controller/src/__tests__/account.test.ts
@@ -1,7 +1,7 @@
 import { hash } from "starknet";
-import { CartridgeSessionAccount } from "../session/ts/account";
-import type { CallPolicy, Session } from "../session/ts/types";
-import { normalizeFelt } from "../session/ts/shared";
+import { CartridgeSessionAccount } from "../session/internal/account";
+import type { CallPolicy, Session } from "../session/internal/types";
+import { normalizeFelt } from "../session/internal/shared";
 
 // Mock global fetch
 const mockFetch = jest.fn();
