+++ packages/controller/src/node/provider.ts
@@ -1,7 +1,7 @@
 import { ec, encode, stark, WalletAccount } from "starknet";
 import { loadConfig, SessionPolicies } from "@cartridge/presets";
 import { AddStarknetChainParameters } from "@starknet-io/types-js";
-import { signerToGuid } from "../session/ts/shared";
+import { signerToGuid } from "../session/internal/shared";
 
 import SessionAccount from "./account";
 import { KEYCHAIN_URL } from "../constants";
