+++ packages/controller/src/session/index.ts
@@ -2,4 +2,4 @@ export { default } from "./provider";
 export * from "./provider";
 export * from "../errors";
 export * from "../types";
-export * from "./ts/types";
+export * from "./internal/types";
